+++ src/KT0803.h
@@ -88,6 +88,17 @@ public:
   //  CHANNEL
   bool     setChannel(uint16_t channel);
   uint16_t getChannel();
+
+  //  NOT IN KT0803
+  bool     setMono();
+  bool     setStereo();
+  bool     isStereo();
+
+  bool     setBass(uint8_t bass);  //  0..3 = 0, 5, 11, 17 dB
+  uint8_t  getBass();
+
+  bool     powerOK();
+  bool     silenceDetected();
 };